Revert "Runelookup to avoid testing regexps that have no chance to match (#111)"
This reverts commit 7621a3216c01965b2ded34f8557fc55062fd4a91.

--

I've reverted this for a couple of reasons. The first is just a bug that
could be fixed, in  that ordering of rules with the same prefix aren't
respected (eg. `${` vs. `$(`).

The second reason is that on reflection I think that Participle will be
better served by keeping the default dynamic implementation fairly
simple and investing most optimisations into code generation. To that
end, while this lookup does bring some improvement (~5%-10% from
memory), it is also quite complex.