More tests and fixes for gen lexer word boundary
Third time's the charm!

The correct word boundary test when the position is NOT at the beginning
and NOT at the end is to always test the character before the current
position and the character at the current position.

In the first version, it tested the character at the current position
and the next character after the current position.

In the second version, it also tested the character
before the current position and the character at the current position.

The correct solution is simpler and makes sense.

Added more conformance tests to cover the changes, including more cases
to ensure \b does NOT match when it should not.