Improve the choice of error reporting after optional branches
The last error in the added test explains the difference. Without this
change, it would have been `1:15: unexpected token "(" (expected ";")`,
which would be very confusing as '(' was actually allowed after the
class name.

After this change, it's `1:20: unexpected token ")" (expected <ident>)`,
pointing out the real problem - that an identifier was expected after
the comma.